+++ R/read_emigration.R
@@ -7,6 +7,7 @@
 #' @template year
 #' @template columns
 #' @template add_labels
+#' @template merge_households
 #' @template as_data_frame
 #' @template showProgress
 #' @template cache
@@ -29,6 +30,7 @@
 read_emigration <- function(year = 2010,
                             columns = NULL,
                             add_labels = NULL,
+                            merge_households = FALSE,
                             as_data_frame = FALSE,
                             showProgress = TRUE,
                             cache = TRUE){
@@ -37,6 +39,7 @@ read_emigration <- function(year = 2010,
   checkmate::assert_numeric(year)
   checkmate::assert_vector(columns, null.ok = TRUE)
   checkmate::assert_logical(as_data_frame)
+  checkmate::assert_logical(merge_households)
   checkmate::assert_string(add_labels, pattern = 'pt', null.ok = TRUE)
 
   # data available for the years:
@@ -60,6 +63,14 @@ read_emigration <- function(year = 2010,
 
   ### read data
   df <- arrow_open_dataset(local_file)
+
+  ### merge household data
+  if (isTRUE(merge_households)) {
+    df <- merge_household_var(df,
+                              year = year,
+                              add_labels = add_labels,
+                              showProgress)
+  }
 
   ### Select
   if (!is.null(columns)) { # columns <- c('V0002','V0011')