+++ man/roxygen/templates/merge_households.R
@@ -1,0 +1,3 @@
+#' @param merge_households Logical. Indicate whether the function should merge
+#'        household variables to the output data. Defaults to `FALSE`.
+